Fixed cpplint conflict and build error

+++ modules/data_structures/source/csr.cpp
@@ -1,6 +1,7 @@
 // Copyright [year] <Copyright Owner>
 #include <vector>
 #include <utility>
+#include <string>
 #include <fstream>
 #include "../include/graph.h"
 #include "../include/csr.h"
@@ -8,8 +9,8 @@
 
 CSR::CSR(const AL& al, ...) {
     offset.push_back(0);
-    for (int i = 0; i < al.edges.size(); ++i) {
-        for (int j = 0; j < al.edges[i].size(); ++j) {
+    for (int i = 0; i < (int)al.edges.size(); ++i) {
+        for (int j = 0; j < (int)al.edges[i].size(); ++j) {
             edges.emplace_back(al.edges[i][j]);
         }
         offset.push_back(edges.size());
